Arreglo de modelos

+++ 20241CYA12A-G2/Models/Descuento.cs
@@ -6,7 +6,7 @@ namespace _20241CYA12A_G2.Models
         public int Id { get; set; }
         public int Dia { get; set; }
         public int Porcentaje { get; set; }
-        public int DescuentoMaximo { get; set; }
+        public decimal DescuentoMaximo { get; set; }
         public bool Activo { get; set; }
 
         //Relaciones
